First implementation of multi-threaded buffering
Buffering now happens concurrently with playing the song. Currently,
there is no mechanism for interrupting a buffering thread if it becomes
necessary to skip ahead quickly. Thus if the user clicks "next" (or
"previous") several times rapidly, the program will get stuck. This must
be taken care of in future commits.